fix(code-block): stop auto-detecting language for empty code fences
`CodeBlockLowlight` falls through to `lowlight.highlightAuto()` when
the language is empty or unregistered. On English prose it picks
VB.NET, whose `'` line-comment syntax makes apostrophes in
contractions ("Don't", "isn't") italicize the rest of the paragraph
via `hljs-comment`. Pass `defaultLanguage: "plaintext"` and switch to
`lowlight(all)` so every dropdown id has a real grammar — 12 ids
(Scala, PowerShell, Dockerfile, Haskell, Elixir, Clojure, Erlang,
OCaml, F#, Dart, MATLAB, LaTeX) were absent from `lowlight(common)`
and would have hit the same auto-detect fallback.

Bundled with:
- Split tiptap.ts (514 lines) into focused modules under the project's
  300-line cap: tiptap (47), nodeView (195), dropdown (290),
  languages (83), icons (17).
- Make the copy button async with explicit error state (red X icon +
  `--error` modifier) — previous code showed success even on clipboard
  rejection or unavailable API.
- i18n the dropdown's "Search languages..." placeholder across all
  10 supported locales.
- Extract `handleSharedDropdownKey` so the search and list keyboard
  handlers stop duplicating ArrowUp/Down/Escape arms.
- Move `plaintext` to top of LANGUAGES so Plain Text still appears
  first in the dropdown for fresh blocks.

Adds 9 tests (5 registry-contract + 4 clipboard-contract); 168/168
green for `src/plugins/codeBlockLineNumbers/`.
