+++ src/plugins/codeBlockLineNumbers/tiptap.ts
@@ -1,514 +1,47 @@
 /**
  * Code Block Extension with Line Numbers
  *
- * Purpose: Extends CodeBlockLowlight with a custom NodeView that renders a line-number
- * gutter, a copy button, and a language selector chip, providing a code-editor-like experience in WYSIWYG.
+ * Purpose: Extends CodeBlockLowlight with a custom NodeView that renders a
+ * line-number gutter, a copy button, and a language selector chip, providing
+ * a code-editor-like experience in WYSIWYG.
+ *
+ * Pipeline: ProseMirror code block node → CodeBlockNodeView (DOM scaffolding,
+ * gutter, copy button, language chip) → LanguageDropdown (dropdown lifecycle,
+ * filtering, keyboard nav).
  *
  * Key decisions:
  *   - Built on CodeBlockLowlight (not raw CodeBlock) for syntax highlighting via lowlight
+ *   - `defaultLanguage: "plaintext"` prevents `lowlight.highlightAuto()` from
+ *     mis-detecting empty-language blocks as VB.NET (which italicizes
+ *     apostrophe contractions via the `hljs-comment` CSS rule)
  *   - Line numbers are rendered in a separate gutter div that updates on every mutation
  *   - Language selector is a floating chip positioned inside the code block wrapper
  *   - Language dropdown uses fixed positioning (popup-host aware) to avoid clipping
- *   - Copy button uses navigator.clipboard API with visual checkmark feedback
- *   - Respects settings.lineNumbers toggle for showing/hiding the gutter
+ *   - Copy button uses navigator.clipboard API with success/error feedback
  *
  * Known limitations:
  *   - Line numbers are recounted on every DOM mutation (no incremental update)
- *   - Language list is hardcoded, not dynamically loaded from lowlight's registry
  *
  * @coordinates-with stores/settingsStore.ts — reads view.lineNumbers setting
- * @coordinates-with sourcePopup — uses getPopupHostForDom/toHostCoordsForDom for dropdown positioning
+ * @coordinates-with plugins/codeBlockLineNumbers/languages — language registry + lowlight setup
+ * @coordinates-with plugins/codeBlockLineNumbers/dropdown — dropdown controller
  * @module plugins/codeBlockLineNumbers/tiptap
  */
 import { CodeBlockLowlight } from "@tiptap/extension-code-block-lowlight";
-import { common, createLowlight } from "lowlight";
-import type { Node as ProseMirrorNode } from "@tiptap/pm/model";
-import type { NodeView, ViewMutationRecord } from "@tiptap/pm/view";
-import type { Editor } from "@tiptap/core";
-import { getPopupHostForDom, toHostCoordsForDom } from "@/plugins/sourcePopup";
-import i18n from "@/i18n";
+import { CodeBlockNodeView } from "./nodeView";
+import { lowlight } from "./languages";
 import "./code-block-line-numbers.css";
 import "./hljs-syntax.css";
 
-const lowlight = createLowlight(common);
-
-/** Lucide copy icon (14×14) */
-const COPY_ICON_SVG = `<svg width="14" height="14" viewBox="0 0 24 24" fill="none" stroke="currentColor" stroke-width="2" stroke-linecap="round" stroke-linejoin="round"><rect x="9" y="9" width="13" height="13" rx="2" ry="2"/><path d="M5 15H4a2 2 0 0 1-2-2V4a2 2 0 0 1 2-2h9a2 2 0 0 1 2 2v1"/></svg>`;
-/** Lucide check icon (14×14) */
-const CHECK_ICON_SVG = `<svg width="14" height="14" viewBox="0 0 24 24" fill="none" stroke="currentColor" stroke-width="2" stroke-linecap="round" stroke-linejoin="round"><polyline points="20 6 9 17 4 12"/></svg>`;
-
 /**
- * Common programming languages with display names.
- * Ordered by popularity/usage.
- */
-const LANGUAGES = [
-  { id: "", name: "Plain Text" },
-  { id: "javascript", name: "JavaScript" },
-  { id: "typescript", name: "TypeScript" },
-  { id: "python", name: "Python" },
-  { id: "java", name: "Java" },
-  { id: "c", name: "C" },
-  { id: "cpp", name: "C++" },
-  { id: "csharp", name: "C#" },
-  { id: "go", name: "Go" },
-  { id: "rust", name: "Rust" },
-  { id: "ruby", name: "Ruby" },
-  { id: "php", name: "PHP" },
-  { id: "swift", name: "Swift" },
-  { id: "kotlin", name: "Kotlin" },
-  { id: "scala", name: "Scala" },
-  { id: "html", name: "HTML" },
-  { id: "css", name: "CSS" },
-  { id: "scss", name: "SCSS" },
-  { id: "json", name: "JSON" },
-  { id: "yaml", name: "YAML" },
-  { id: "xml", name: "XML" },
-  { id: "markdown", name: "Markdown" },
-  { id: "sql", name: "SQL" },
-  { id: "shell", name: "Shell" },
-  { id: "bash", name: "Bash" },
-  { id: "powershell", name: "PowerShell" },
-  { id: "dockerfile", name: "Dockerfile" },
-  { id: "graphql", name: "GraphQL" },
-  { id: "lua", name: "Lua" },
-  { id: "r", name: "R" },
-  { id: "perl", name: "Perl" },
-  { id: "haskell", name: "Haskell" },
-  { id: "elixir", name: "Elixir" },
-  { id: "clojure", name: "Clojure" },
-  { id: "erlang", name: "Erlang" },
-  { id: "ocaml", name: "OCaml" },
-  { id: "fsharp", name: "F#" },
-  { id: "dart", name: "Dart" },
-  { id: "objectivec", name: "Objective-C" },
-  { id: "matlab", name: "MATLAB" },
-  { id: "latex", name: "LaTeX" },
-  { id: "diff", name: "Diff" },
-  { id: "plaintext", name: "Plain Text" },
-];
-
-/**
- * Custom NodeView for code blocks with line numbers and language selector.
- */
-class CodeBlockNodeView implements NodeView {
-  dom: HTMLElement;
-  contentDOM: HTMLElement;
-  private gutter: HTMLElement;
-  private codeElement: HTMLElement;
-  private langSelector: HTMLElement;
-  private copyBtn: HTMLElement;
-  private actionsContainer: HTMLElement;
-  private dropdown: HTMLElement | null = null;
-  private dropdownHost: HTMLElement | null = null;
-  private node: ProseMirrorNode;
-  private editor: Editor;
-  private getPos: () => number | undefined;
-
-  constructor(node: ProseMirrorNode, editor: Editor, getPos: () => number | undefined) {
-    this.node = node;
-    this.editor = editor;
-    this.getPos = getPos;
-
-    // Create wrapper
-    this.dom = document.createElement("div");
-    this.dom.className = "code-block-wrapper";
-
-    // Create gutter for line numbers
-    this.gutter = document.createElement("div");
-    this.gutter.className = "code-line-numbers";
-    this.gutter.setAttribute("aria-hidden", "true");
-    this.gutter.contentEditable = "false";
-    this.dom.appendChild(this.gutter);
-
-    // Create pre element
-    const pre = document.createElement("pre");
-    this.dom.appendChild(pre);
-
-    // Create code element (contentDOM)
-    this.codeElement = document.createElement("code");
-    if (node.attrs.language) {
-      this.codeElement.className = `language-${node.attrs.language}`;
-    }
-    pre.appendChild(this.codeElement);
-    this.contentDOM = this.codeElement;
-
-    // Create language selector
-    this.langSelector = document.createElement("div");
-    this.langSelector.className = "code-lang-selector";
-    this.langSelector.contentEditable = "false";
-    this.updateLangSelectorText();
-    // Use mousedown with capture phase to get event before ProseMirror
-    this.langSelector.addEventListener("mousedown", this.handleLangClick, { capture: true });
-
-    // Create copy button
-    this.copyBtn = document.createElement("button");
-    this.copyBtn.className = "code-copy-btn";
-    this.copyBtn.innerHTML = COPY_ICON_SVG;
-    const copyLabel = i18n.t("editor:plugin.copySource");
-    this.copyBtn.title = copyLabel;
-    this.copyBtn.setAttribute("aria-label", copyLabel);
-    this.copyBtn.addEventListener("mousedown", this.handleCopyMouseDown);
-    this.copyBtn.addEventListener("click", this.handleCopyClick);
-
-    // Actions container: copy button + language selector
-    this.actionsContainer = document.createElement("div");
-    this.actionsContainer.className = "code-block-actions";
-    this.actionsContainer.contentEditable = "false";
-    this.actionsContainer.appendChild(this.copyBtn);
-    this.actionsContainer.appendChild(this.langSelector);
-    this.dom.appendChild(this.actionsContainer);
-
-    // Initial line count
-    this.updateLineNumbers();
-  }
-
-  update(node: ProseMirrorNode): boolean {
-    if (node.type !== this.node.type) return false;
-    this.node = node;
-
-    // Update language class
-    if (node.attrs.language) {
-      this.codeElement.className = `language-${node.attrs.language}`;
-    } else {
-      this.codeElement.className = "";
-    }
-
-    // Update language selector text
-    this.updateLangSelectorText();
-
-    // Update line numbers
-    this.updateLineNumbers();
-    return true;
-  }
-
-  destroy(): void {
-    this.closeDropdown();
-    this.langSelector.removeEventListener("mousedown", this.handleLangClick, { capture: true });
-    this.copyBtn.removeEventListener("mousedown", this.handleCopyMouseDown);
-    this.copyBtn.removeEventListener("click", this.handleCopyClick);
-  }
-
-  private updateLangSelectorText(): void {
-    const lang = this.node.attrs.language || "";
-    const langInfo = LANGUAGES.find((l) => l.id === lang);
-    /* v8 ignore next -- @preserve The LANGUAGES list includes an entry with id="" (name "Plain Text"), so langInfo is always found when lang is empty; the || "Plain Text" fallback is unreachable with the current language list */
-    this.langSelector.textContent = langInfo?.name || lang || "Plain Text";
-  }
-
-  private updateLineNumbers(): void {
-    const text = this.node.textContent;
-    const lineCount = text.split("\n").length;
-
-    // Clear existing line numbers
-    this.gutter.innerHTML = "";
-
-    // Generate line number elements
-    for (let i = 1; i <= lineCount; i++) {
-      const lineNum = document.createElement("div");
-      lineNum.className = "line-num";
-      lineNum.textContent = String(i);
-      this.gutter.appendChild(lineNum);
-    }
-  }
-
-  private handleCopyMouseDown = (e: MouseEvent): void => {
-    e.preventDefault();
-    e.stopPropagation();
-  };
-
-  private handleCopyClick = (e: MouseEvent): void => {
-    e.preventDefault();
-    e.stopPropagation();
-
-    navigator.clipboard?.writeText(this.node.textContent);
-    this.copyBtn.innerHTML = CHECK_ICON_SVG;
-    this.copyBtn.classList.add("code-copy-btn--success");
-    setTimeout(() => {
-      this.copyBtn.innerHTML = COPY_ICON_SVG;
-      this.copyBtn.classList.remove("code-copy-btn--success");
-    }, 1500);
-  };
-
-  private handleLangClick = (e: MouseEvent): void => {
-    e.preventDefault();
-    e.stopPropagation();
-    e.stopImmediatePropagation();
-
-    if (this.dropdown) {
-      this.closeDropdown();
-    } else {
-      this.openDropdown();
-    }
-  };
-
-  private openDropdown(): void {
-    /* v8 ignore next -- @preserve Defensive guard: handleLangClick calls closeDropdown when dropdown exists, so openDropdown is only called when dropdown is null; this guard prevents double-open if called directly */
-    if (this.dropdown) return;
-
-    const dropdown = document.createElement("div");
-    dropdown.className = "code-lang-dropdown";
-
-    // Search input
-    const searchInput = document.createElement("input");
-    searchInput.type = "text";
-    searchInput.className = "code-lang-search";
-    searchInput.placeholder = "Search languages...";
-    searchInput.addEventListener("input", () => this.filterLanguages(searchInput.value));
-    searchInput.addEventListener("keydown", (e) => this.handleSearchKeydown(e));
-    dropdown.appendChild(searchInput);
-
-    // Language list
-    const list = document.createElement("div");
-    list.className = "code-lang-list";
-    this.renderLanguageList(list, "");
-    dropdown.appendChild(list);
-
-    this.dropdown = dropdown;
-
-    // Mount to editor container if available, otherwise document.body
-    this.dropdownHost = getPopupHostForDom(this.dom) ?? document.body;
-    dropdown.style.position = this.dropdownHost === document.body ? "fixed" : "absolute";
-    this.dropdownHost.appendChild(dropdown);
-    this.positionDropdown();
-
-    // Focus search input
-    requestAnimationFrame(() => searchInput.focus());
-
-    // Close on outside click
-    document.addEventListener("mousedown", this.handleOutsideClick);
-    // Reposition on scroll
-    window.addEventListener("scroll", this.positionDropdown, true);
-  }
-
-  private positionDropdown = (): void => {
-    /* v8 ignore next -- @preserve Defensive guard: scroll listener is removed by closeDropdown before dropdown is nulled, so positionDropdown is only called when dropdown exists */
-    if (!this.dropdown) return;
-    const rect = this.langSelector.getBoundingClientRect();
-    const top = rect.bottom + 4;
-    const left = rect.right - 180; // align right edge
-
-    // Convert to host-relative coordinates if mounted inside editor container
-    if (this.dropdownHost !== document.body && this.dropdownHost) {
-      const hostPos = toHostCoordsForDom(this.dropdownHost, { top, left });
-      this.dropdown.style.top = `${hostPos.top}px`;
-      this.dropdown.style.left = `${hostPos.left}px`;
-    } else {
-      this.dropdown.style.top = `${top}px`;
-      this.dropdown.style.left = `${left}px`;
-    }
-  };
-
-  private closeDropdown(): void {
-    if (this.dropdown) {
-      this.dropdown.remove();
-      this.dropdown = null;
-      this.dropdownHost = null;
-      document.removeEventListener("mousedown", this.handleOutsideClick);
-      window.removeEventListener("scroll", this.positionDropdown, true);
-    }
-  }
-
-  private handleOutsideClick = (e: MouseEvent): void => {
-    if (this.dropdown && !this.dropdown.contains(e.target as Node) && !this.langSelector.contains(e.target as Node)) {
-      this.closeDropdown();
-    }
-  };
-
-  private filterLanguages(query: string): void {
-    if (!this.dropdown) return;
-    const list = this.dropdown.querySelector(".code-lang-list");
-    if (list) {
-      this.renderLanguageList(list as HTMLElement, query);
-    }
-  }
-
-  private renderLanguageList(container: HTMLElement, query: string): void {
-    container.innerHTML = "";
-    const lowerQuery = query.toLowerCase();
-    const filtered = LANGUAGES.filter(
-      (lang) => lang.name.toLowerCase().includes(lowerQuery) || lang.id.toLowerCase().includes(lowerQuery)
-    );
-
-    const currentLang = this.node.attrs.language || "";
-    // Find index of current language in filtered list
-    const currentIndex = filtered.findIndex((lang) => lang.id === currentLang);
-    // Highlight current language if found, otherwise first item
-    const highlightIndex = currentIndex >= 0 ? currentIndex : 0;
-
-    filtered.forEach((lang, index) => {
-      const item = document.createElement("div");
-      item.className = "code-lang-item";
-      item.tabIndex = 0; // Make focusable
-      if (lang.id === currentLang) {
-        item.classList.add("active");
-      }
-      if (index === highlightIndex) {
-        item.classList.add("highlighted");
-      }
-      item.textContent = lang.name;
-      item.dataset.langId = lang.id;
-      item.addEventListener("click", () => this.selectLanguage(lang.id));
-      item.addEventListener("keydown", this.handleListKeydown);
-      container.appendChild(item);
-    });
-
-    // Scroll highlighted item into view
-    requestAnimationFrame(() => {
-      const highlighted = container.querySelector(".highlighted");
-      if (highlighted) {
-        highlighted.scrollIntoView({ block: "nearest" });
-      }
-    });
-  }
-
-  private handleSearchKeydown = (e: KeyboardEvent): void => {
-    if (!this.dropdown) return;
-
-    const list = this.dropdown.querySelector(".code-lang-list");
-    if (!list) return;
-
-    const items = Array.from(list.querySelectorAll(".code-lang-item")) as HTMLElement[];
-    if (items.length === 0) return;
-
-    switch (e.key) {
-      case "Tab": {
-        // Tab moves focus to the highlighted item in the list
-        e.preventDefault();
-        const highlighted = list.querySelector(".code-lang-item.highlighted") as HTMLElement;
-        /* v8 ignore next -- @preserve reason: false branch (no highlighted item) falls to defensive items[0] guard */
-        if (highlighted) {
-          highlighted.focus();
-        /* v8 ignore start -- @preserve items[0] is always truthy here: items.length === 0 guard above ensures non-empty */
-        } else if (items[0]) {
-          items[0].classList.add("highlighted");
-          items[0].focus();
-        }
-        /* v8 ignore stop */
-        break;
-      }
-      case "ArrowDown":
-        e.preventDefault();
-        this.moveHighlight(items, 1);
-        break;
-      case "ArrowUp":
-        e.preventDefault();
-        this.moveHighlight(items, -1);
-        break;
-      case "Enter": {
-        e.preventDefault();
-        const current = list.querySelector(".code-lang-item.highlighted") as HTMLElement;
-        if (current) {
-          const langId = current.dataset.langId || "";
-          this.selectLanguage(langId);
-        }
-        break;
-      }
-      case "Escape":
-        e.preventDefault();
-        this.closeDropdown();
-        break;
-    }
-  };
-
-  private handleListKeydown = (e: KeyboardEvent): void => {
-    if (!this.dropdown) return;
-
-    const list = this.dropdown.querySelector(".code-lang-list");
-    if (!list) return;
-
-    const items = Array.from(list.querySelectorAll(".code-lang-item")) as HTMLElement[];
-    if (items.length === 0) return;
-
-    switch (e.key) {
-      case "ArrowDown":
-        e.preventDefault();
-        this.moveHighlight(items, 1);
-        break;
-      case "ArrowUp":
-        e.preventDefault();
-        this.moveHighlight(items, -1);
-        break;
-      case "Tab": {
-        // Shift+Tab goes back to search input
-        if (e.shiftKey) {
-          e.preventDefault();
-          const searchInput = this.dropdown.querySelector(".code-lang-search") as HTMLInputElement;
-          if (searchInput) {
-            searchInput.focus();
-          }
-        } else {
-          // Tab without shift moves to next item
-          e.preventDefault();
-          this.moveHighlight(items, 1);
-        }
-        break;
-      }
-      case "Enter": {
-        e.preventDefault();
-        const target = e.target as HTMLElement;
-        if (target.classList.contains("code-lang-item")) {
-          const langId = target.dataset.langId || "";
-          this.selectLanguage(langId);
-        }
-        break;
-      }
-      case "Escape":
-        e.preventDefault();
-        this.closeDropdown();
-        break;
-    }
-  };
-
-  private moveHighlight(items: HTMLElement[], direction: number): void {
-    const currentHighlighted = items.find((item) => item.classList.contains("highlighted"));
-    const currentIndex = currentHighlighted ? items.indexOf(currentHighlighted) : -1;
-
-    // Remove current highlight
-    currentHighlighted?.classList.remove("highlighted");
-
-    // Calculate new index
-    let newIndex = currentIndex + direction;
-    if (newIndex < 0) newIndex = 0;
-    if (newIndex >= items.length) newIndex = items.length - 1;
-
-    // Apply new highlight and focus
-    items[newIndex].classList.add("highlighted");
-    items[newIndex].scrollIntoView({ block: "nearest" });
-    items[newIndex].focus();
-  }
-
-  private selectLanguage(langId: string): void {
-    const pos = this.getPos();
-    if (pos === undefined) return;
-
-    this.editor.chain().focus().updateAttributes("codeBlock", { language: langId }).run();
-    this.closeDropdown();
-  }
-
-  ignoreMutation(mutation: ViewMutationRecord): boolean {
-    // Ignore mutations to the gutter and language selector
-    if (mutation.type === "selection") {
-      return false;
-    }
-    if (this.gutter.contains(mutation.target as Node)) {
-      return true;
-    }
-    if (this.actionsContainer.contains(mutation.target as Node)) {
-      return true;
-    }
-    if (this.dropdown?.contains(mutation.target as Node)) {
-      return true;
-    }
-    return false;
-  }
-}
-
-/**
- * Extended CodeBlockLowlight with line numbers and language selector.
+ * Extended CodeBlockLowlight with line numbers, a copy button, and a language
+ * selector dropdown. `defaultLanguage: "plaintext"` is load-bearing — without
+ * it, `@tiptap/extension-code-block-lowlight` invokes `lowlight.highlightAuto`
+ * on empty-language blocks and frequently mis-detects English prose as VB.NET.
  */
 export const CodeBlockWithLineNumbers = CodeBlockLowlight.extend({
   addNodeView() {
-    return ({ node, editor, getPos }) => new CodeBlockNodeView(node, editor, getPos as () => number | undefined);
+    return ({ node, editor, getPos }) =>
+      new CodeBlockNodeView(node, editor, getPos as () => number | undefined);
   },
-
-}).configure({ lowlight });
+}).configure({ lowlight, defaultLanguage: "plaintext" });
